+++ lib/app_with_flutter_bloc/features/o_b_s_sources/repositories/sources_repository.dart
@@ -1,5 +1,5 @@
-import 'package:obs_for_sama/app_with_flutter_bloc/features/server/exceptions/server_exception.dart';
 import 'package:obs_for_sama/app_with_flutter_bloc/features/server/singleton/o_b_s_singleton.dart';
+import 'package:obs_for_sama/core/exceptions/exceptions.dart';
 import 'package:obs_websocket/event.dart';
 import 'package:obs_websocket/obs_websocket.dart';
 
@@ -8,16 +8,20 @@ class SourcesRepository {
     required SceneItemDetail source,
     required String currentSceneName,
   }) async {
-    final ObsWebSocket? obsWebSocket = await OBSSingleton().obs;
+    try {
+      final ObsWebSocket? obsWebSocket = await OBSSingleton().obs;
 
-    final SceneItemEnableStateChanged sceneItemEnableStateChanged = SceneItemEnableStateChanged(
-      sceneName: currentSceneName,
-      sceneItemId: source.sceneItemId,
-      sceneItemEnabled: !source.sceneItemEnabled,
-    );
+      final SceneItemEnableStateChanged sceneItemEnableStateChanged = SceneItemEnableStateChanged(
+        sceneName: currentSceneName,
+        sceneItemId: source.sceneItemId,
+        sceneItemEnabled: !source.sceneItemEnabled,
+      );
 
-    await obsWebSocket?.sceneItems.setSceneItemEnabled(sceneItemEnableStateChanged);
-    return getListSourcesByCurrentScene(currentSceneName: currentSceneName);
+      await obsWebSocket?.sceneItems.setSceneItemEnabled(sceneItemEnableStateChanged);
+      return getListSourcesByCurrentScene(currentSceneName: currentSceneName);
+    } on Exception {
+      throw OBSSourcesException('SOURCES_TOOGLE_ERROR');
+    }
   }
 
   static Future<List<SceneItemDetail>> getListSourcesByCurrentScene({required String currentSceneName}) async {
@@ -26,8 +30,8 @@ class SourcesRepository {
 
       final List<SceneItemDetail> details = await obsWebSocket?.sceneItems.getSceneItemList(currentSceneName) ?? [];
       return details.reversed.toList();
-    } on Exception catch (_) {
-      throw ServerException('SOURCES_ERROR');
+    } on Exception {
+      throw OBSSourcesException('SOURCES_LIST_ERROR');
     }
   }
 }