+++ lib/app_with_flutter_bloc/features/server/exceptions/server_exception.dart
@@ -1,7 +1,0 @@
-class ServerException implements Exception {
-  ServerException(this.message);
-  final String message;
-
-  @override
-  String toString() => '\n----------------\nClass : [ServerException]\nMessage : $message\n----------------\n';
-}